Add a BOLT11 parser using Browserify
A BOLT11 parser released as an npm module has been added using browserify, in order to allow to pay BOLT11 payment requests through a future update.

Since the module uses some ES6 rules, a transpiler is also necessary to improve the compatibility with older browsers and to avoid breaking the usemin task which runs during the building process.